refactor: Start Image cleanup
