+++ src/Uno.UI/UI/Xaml/Controls/Image/Legacy/Image.shared.cs
@@ -8,12 +8,6 @@ partial class Image
 	private Color? _monochromeColor;
 
 	/// <summary>
-	/// Creates and returns a ImageAutomationPeer object for this Image.
-	/// </summary>
-	/// <returns>ImageAutomationPeer.</returns>
-	protected override AutomationPeer OnCreateAutomationPeer() => new ImageAutomationPeer(this);
-
-	/// <summary>
 	/// When set, the resulting image is tentatively converted to Monochrome.
 	/// </summary>
 	internal Color? MonochromeColor
