+++ src/Uno.UI/UI/Xaml/Controls/Image/Legacy/Image.svg.cs
