+++ src/Uno.UI/UI/Xaml/Controls/Image/Native/HtmlImage.wasm.cs
